Convert lines to ints and return to main func
Needed to strip the newlines from each line read in and then
convert to integer.  It made sense to pack all of these values
into an array and return them as such.  To prove this works
I have the main func output them to stdout.

Note, I commented out the fuel calculation statement and changed
the default return value.  Will revisit this in a follow-on commit.