Now can calculate fuel's fuel cost
Implemented the cost calculation recursively to make use of the same
function.  The input mass module values are never greater than six
digits so there is little concern of running out of memory.

Along with the reimplementation, I updated the test cases to reflect
the new calculation.  I ran these manually to confirm but there should
be some automation in place for gating.